Add python interface for distance to closest fault Split up some older code into multiple files.
