+++ GrpcInterface/GrpcProtos/Case.proto
@@ -21,6 +21,7 @@ service Case {
   rpc GetCaseInfo(CaseRequest) returns (CaseInfo) {}
   rpc GetPdmObject(CaseRequest) returns (PdmObject) {}
   rpc GetReservoirBoundingBox(CaseRequest) returns (BoundingBox) {}
+  rpc GetDistanceToClosestFault(ClosestFaultRequest) returns (ClosestFault) {}
 }
 
 message CaseRequest { int32 id = 1; }
@@ -103,3 +104,14 @@ message SelectedCell {
 }
 
 message SelectedCells { repeated SelectedCell cells = 1; }
+
+message ClosestFaultRequest {
+  CaseRequest case_request = 1;
+  Vec3d point = 2;
+}
+
+message ClosestFault {
+    string fault_name = 1;
+    double distance = 2;
+    string face_name = 3;
+}
